requirements.txt
cycler==0.10.0
matplotlib==2.1.1
nose==1.3.7
numpy==1.14.0
Pillow==5.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
